deepl/code/requirements.txt
deepl
tqdm
python-dotenv